+++ scripts/benchmark_latency_bandwidth.sh
@@ -1,13 +1,15 @@
 #!/bin/bash
 
-# Benchmark a communication pattern for various world sizes and dtypes
-# Generates multiple CSV files
-
+# Benchmark a communication pattern for various dtypes.
+# Generates multiple CSV/PNG files
 
 
 export DEVICE="cpu"
 DTYPES="int float16 float32"
 DTYPES="float16"
+
+# To surpress a torchrun warning
+export OMP_NUM_THREADS=1
 
 BASE_DIR=$(dirname "$0")
 
@@ -20,7 +22,10 @@ do
     do
         export HCCL_DTYPE=${DTYPE}
         echo "${BENCH} (size: ${SIZE} | dtype: ${DTYPE})"
-        make "${BENCH}" -e DEVICE="${DEVICE}" HCCL_DTYPE="${HCCL_DTYPE}"
+        torchrun --nnodes 1 --nproc_per_node 2 pytorch_hccl_tests/cli.py \
+                    --benchmark "${BENCH}" \
+                    --device ${DEVICE} \
+                    --dtype "${HCCL_DTYPE}"
 
         python "${BASE_DIR}/plotter_${BENCH}.py"
     done