+++ scripts/plotter_collectives.py
@@ -74,7 +74,7 @@ def main():
     ax.set(yscale="log")
     ax.legend(markerscale=2)
     ax.set_xlabel("Message length (bytes)")
-    ax.set_ylabel("Average Latency (us)")
+    ax.set_ylabel("Average Elapsed Time (ns)")
     title = f"OSU-MPI {BENCHMARK} benchmark\n (Device: {DEVICE} | dtype: {DTYPE}"
     title += f" | PT: {PT_VER}"
     title += ")"